Refactoring to allow splitting of installation scripts
