+++ offworld_gym/envs/gazebo/utils.py
@@ -18,43 +18,11 @@ __version__     = version.__version__
 
 import numpy as np
 import cv2
-
 import rospy
 from std_srvs.srv import Empty as Empty_srv
 from cv_bridge import CvBridge, CvBridgeError
 from gazebo_msgs.srv import GetModelState, SetModelState
 from gazebo_msgs.msg import ModelState
-
-class ImageUtils(object):
-    """Image utility functions used by the OffWorld Gym environments
-    """
-    IMG_H = 240
-    IMG_W = 320
-    IMG_C = 3
-
-    @staticmethod
-    def process_img_msg(img_msg, resized_width=IMG_W, resized_height=IMG_H):
-        """Converts ROS image to cv2, then to numpy
-        """
-        img = CvBridge().imgmsg_to_cv2(img_msg, "bgr8")
-        img = np.asarray(img)
-        img = cv2.cvtColor(img, cv2.COLOR_BGR2RGB)
-        img = cv2.resize(img, (resized_width, resized_height))
-
-        img = np.reshape(img, (1, img.shape[0], img.shape[1], img.shape[2]))
-        return img 
-    
-    @staticmethod
-    def process_depth_msg(depth_msg, resized_width=IMG_W, resized_height=IMG_H):
-        """Converts a depth image into numpy float32 array
-        """
-        cv_image = CvBridge().imgmsg_to_cv2(depth_msg, "32FC1")
-        img = np.asarray(cv_image, dtype=np.float32)
-        img = np.nan_to_num(img)
-        img = cv2.resize(img, (resized_width, resized_height))
-
-        img = np.reshape(img, (1, img.shape[0], img.shape[1], 1))
-        return img
 
 class GazeboUtils:
     """Gazebo utility functions used by the OffWorld Gym environments
@@ -100,3 +68,35 @@ class GazeboUtils:
         except rospy.ServiceException as e:
             rospy.logerr('The robot cannot be reset.')
             raise e
+
+class ImageUtils(object):
+    """Image utility functions used by the OffWorld Gym environments
+    """
+    IMG_H = 240
+    IMG_W = 320
+    IMG_C = 3
+
+    @staticmethod
+    def process_img_msg(img_msg, resized_width=IMG_W, resized_height=IMG_H):
+        """Converts ROS image to cv2, then to numpy
+        """
+        img = CvBridge().imgmsg_to_cv2(img_msg, "bgr8")
+        img = np.asarray(img)
+        img = cv2.cvtColor(img, cv2.COLOR_BGR2RGB)
+        img = cv2.resize(img, (resized_width, resized_height))
+
+        img = np.reshape(img, (1, img.shape[0], img.shape[1], img.shape[2]))
+        return img 
+    
+    @staticmethod
+    def process_depth_msg(depth_msg, resized_width=IMG_W, resized_height=IMG_H):
+        """Converts a depth image into numpy float32 array
+        """
+        cv_image = CvBridge().imgmsg_to_cv2(depth_msg, "32FC1")
+        img = np.asarray(cv_image, dtype=np.float32)
+        img = np.nan_to_num(img)
+        img = cv2.resize(img, (resized_width, resized_height))
+
+        img = np.reshape(img, (1, img.shape[0], img.shape[1], 1))
+        return img
+
